feat: worked on comments
